XP-1292 Move LocalizeParams into portal-impl
- Moved LocalizeParams into portal-impl (com.enonic.xp.portal.impl.view) as a package protected class.
