+++ modules/portal-impl/src/main/java/com/enonic/xp/portal/impl/view/LocalizeFunction.java
@@ -8,7 +8,6 @@ import org.osgi.service.component.annotations.Reference;
 import com.enonic.xp.i18n.LocaleService;
 import com.enonic.xp.i18n.MessageBundle;
 import com.enonic.xp.portal.PortalRequestAccessor;
-import com.enonic.xp.portal.i18n.LocalizeParams;
 import com.enonic.xp.portal.view.ViewFunction;
 import com.enonic.xp.portal.view.ViewFunctionParams;
